requirements.txt
flask
flask_cors
jinja2
googleapis_common_protos
google-cloud-ndb
google-cloud-logging
google-cloud-tasks
google-cloud-datastore
